+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # AdGuard Home Raycast Extension Changelog
 
+## [1.1.0] - 2024-03-20
+
+- Added snooze protection feature with multiple duration options
+- Added countdown timer for snooze status
+- Added ability to snooze until tomorrow
+- Improved status display for protection state
+
 ## [1.0.0] - 2024-03-19
 
 - Toggle AdGuard Home protection